Mark panic_handler! macro invocation as unknown-os (#93)
ASR provides a replacement for the standard library's Instant under
time_util::Instant. This requires a WASI target. When targeting WASI,
the default panic handler implementation conflicts. I don't intend to
make any features that actually require WASI in the main branch, but
it's useful for ad-hoc time measuring, and this means I don't need to
comment out the panic handler line when switching to WASI and uncomment
it when removing the timestamp code.